Fixes for -Wall warnings

+++ coda-src/sidekick/adv_sidekick.cc
@@ -305,7 +305,7 @@ int executor(char *pathname, int vuid, int req_no)
 		if (!(sbuf.st_mode & S_IFREG))
 		    quit("File %s is not a regular file (and hence cannot be used for repair)", fixed);
 
-		if (!repair_getfid(fixed, &fixfid, fixrealm, &fixvv, space, sizeof(space)) && (fixvv.StoreId.Host != -1))
+		if (!repair_getfid(fixed, &fixfid, fixrealm, &fixvv, space, sizeof(space)) && (fixvv.StoreId.Host != (unsigned long)-1))
 		    sprintf(fixfile, "@%lx.%lx.%lx@%s", fixfid.Volume, fixfid.Vnode, fixfid.Unique, fixrealm);
 		else strcpy(fixfile, fixed);
